+++ airnowpy/observation.py
@@ -12,7 +12,7 @@ class Observation(object):
     """
 
     def __init__(self,
-            timestampUTC: datetime,
+            timestamp: datetime,
             reportingArea: str,
             stateCode: str,
             latitude: float,
@@ -22,7 +22,7 @@ class Observation(object):
             category: Category):
         """
         Parameters:
-            timestampUTC (datetime): The UTC timestamp of the observation
+            timestamp (datetime): The local timestamp of the observation
             reportingArea (string):	City or area name of observed data
             stateCode (string): Two-character state abbreviation
             latitude (float): Latitude in decimal degrees
@@ -32,7 +32,7 @@ class Observation(object):
             category (Category): The corresponding Category for the AQI value
         """
 
-        self._timestampUTC = timestampUTC
+        self._timestamp = timestamp
         self._reportingArea = reportingArea
         self._stateCode = stateCode
         self._latitude = latitude
@@ -47,5 +47,5 @@ class Observation(object):
     def getCategory(self) -> Category:
         return self._category
 
-    def getTimestampUTC(self) -> datetime:
-        return self._timestampUTC
+    def getTimestamp(self) -> datetime:
+        return self._timestamp